requirements.txt
pytorch==2.1.0
transformers=4.34.1
bitsandbytes=0.41.1
peft=0.5.0
datasets=2.14.5
evaluate
sentencepiece
gpustat
wandb
bigmodelvis
matplotlib
scikit-learn
accelerate
numpy
pandas
scipy
pyarrow==12.0.1
promptsource
gpustat